Swap out all email links with <LinkTypeWrapper>
- update tests
